edit some formatting issue
