+++ src/contactus/body/Form.js
@@ -6,9 +6,11 @@ import { Formik } from 'formik';
 import * as yup from 'yup';
 import emailjs from 'emailjs-com';
 import FormHolder from './FormHolder';
+// import Notification from './Notification';
 
 const Form = () => {
   const [type, setType] = useState('Type (Optional)');
+  const [isShowMessage, setIsShowMessage] = useState(false);
 
   const listOfTypes = [
     { type: 'Type (Optional)', key: 0 },
@@ -66,75 +68,80 @@ const Form = () => {
       onSubmit={(values, actions) => {
         sendMessage(values);
         setType('Type (Optional)');
+        setIsShowMessage(true);
+        setTimeout(() => setIsShowMessage(false), 5000);
         actions.resetForm();
       }}
     >
       {(props) => (
-        <div className="form-box">
-          <div className="form-particulars" id="form-particulars">
-            <FormHolder classNameToUse="form-particular-error-holder">
-              <div className="form-particulars-holder">
-                <input
-                  placeholder="Name"
-                  required={true}
-                  onChange={props.handleChange}
-                  value={props.values.name}
-                  name="name"
-                />
+        <>
+          <div className="form-box">
+            <div className="form-particulars" id="form-particulars">
+              <FormHolder classNameToUse="form-particular-error-holder">
+                <div className="form-particulars-holder">
+                  <input
+                    placeholder="Name"
+                    required={true}
+                    onChange={props.handleChange}
+                    value={props.values.name}
+                    name="name"
+                  />
+                </div>
+                <span>{props.touched.name && props.errors.name && 'Name is a required field'}</span>
+              </FormHolder>
+              <FormHolder classNameToUse="form-particular-error-holder">
+                <div className="form-particulars-holder">
+                  <input
+                    placeholder="Email"
+                    required={true}
+                    onChange={props.handleChange}
+                    value={props.values.email}
+                    name="email"
+                  />
+                </div>
+                <span>
+                  {props.touched.email && props.errors.email && 'Email is a required field'}
+                </span>
+              </FormHolder>
+              <FormHolder classNameToUse="form-particular-error-holder">
+                <div className="form-particulars-holder">
+                  <span>{type}</span>
+                  <Dropdown overlay={menu} trigger={['click']} placement={'bottomCenter'}>
+                    <a
+                      className="ant-dropdown-link"
+                      onClick={(e) => e.preventDefault()}
+                      id="drop-down-arrow-holder"
+                    >
+                      <DownOutlined style={{ transform: 'translateX(-20px)' }} />
+                    </a>
+                  </Dropdown>
+                </div>
+              </FormHolder>
+            </div>
+            <FormHolder classNameToUse="form-details">
+              <div className="form-details-error-holder">
+                <div className="form-details-holder">
+                  <textarea
+                    placeholder="Message"
+                    required={true}
+                    onChange={props.handleChange}
+                    value={props.values.message}
+                    name="message"
+                  />
+                </div>
+                <span>
+                  {props.touched.message && props.errors.message && 'Message is a required field'}
+                </span>
               </div>
-              <span>{props.touched.name && props.errors.name && 'Name is a required field'}</span>
             </FormHolder>
-            <FormHolder classNameToUse="form-particular-error-holder">
-              <div className="form-particulars-holder">
-                <input
-                  placeholder="Email"
-                  required={true}
-                  onChange={props.handleChange}
-                  value={props.values.email}
-                  name="email"
-                />
-              </div>
-              <span>
-                {props.touched.email && props.errors.email && 'Email is a required field'}
-              </span>
-            </FormHolder>
-            <FormHolder classNameToUse="form-particular-error-holder">
-              <div className="form-particulars-holder">
-                <span>{type}</span>
-                <Dropdown overlay={menu} trigger={['click']} placement={'bottomCenter'}>
-                  <a
-                    className="ant-dropdown-link"
-                    onClick={(e) => e.preventDefault()}
-                    id="drop-down-arrow-holder"
-                  >
-                    <DownOutlined style={{ transform: 'translateX(-20px)' }} />
-                  </a>
-                </Dropdown>
-              </div>
+            <FormHolder classNameToUse="form-submit-button-box">
+              <button className="form-submit-button" onClick={props.handleSubmit} type="submit">
+                <span>SEND MESSAGE</span>
+              </button>
             </FormHolder>
           </div>
-          <FormHolder classNameToUse="form-details">
-            <div className="form-details-error-holder">
-              <div className="form-details-holder">
-                <textarea
-                  placeholder="Message"
-                  required={true}
-                  onChange={props.handleChange}
-                  value={props.values.message}
-                  name="message"
-                />
-              </div>
-              <span>
-                {props.touched.message && props.errors.message && 'Message is a required field'}
-              </span>
-            </div>
-          </FormHolder>
-          <FormHolder classNameToUse="form-submit-button-box">
-            <button className="form-submit-button" onClick={props.handleSubmit} type="submit">
-              <span>SEND MESSAGE</span>
-            </button>
-          </FormHolder>
-        </div>
+          {/*<Notification />*/}
+        </>
       )}
     </Formik>
   );
