+++ src/contactus/body/Notification.js
@@ -1,0 +1,10 @@
+// import { Snackbar } from '@material-ui/core';
+// import { Alert } from '@material-ui/lab';
+//
+// import React from 'react';
+//
+// const Notification = () => {
+//   return <div></div>;
+// };
+//
+// export default Notification;
